Workaround for searching with stopword
