+++ packages/orama/tests/main.test.ts
@@ -214,160 +214,6 @@ t.test('document validation', t => {
 })
 
 t.test('orama', t => {
-  t.test('should correctly search for data', async t => {
-    t.plan(8)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-      components: {
-        tokenizer: {
-          stemming: true,
-        },
-      },
-    })
-
-    await insert(db, { quote: 'the quick, brown fox jumps over the lazy dog. What a fox!', author: 'John Doe' })
-    await insert(db, { quote: 'Foxes are nice animals. But I prefer having a dog.', author: 'John Doe' })
-    await insert(db, { quote: 'I like dogs. They are the best.', author: 'Jane Doe' })
-    await insert(db, { quote: 'I like cats. They are the best.', author: 'Jane Doe' })
-
-    // Exact search
-    const result1 = await search(db, { term: 'fox', exact: true })
-    const result2 = await search(db, { term: 'dog', exact: true })
-
-    t.equal(result1.count, 2)
-    t.equal(result2.count, 3)
-
-    // Prefix search
-    const result3 = await search(db, { term: 'fox', exact: false })
-    const result4 = await search(db, { term: 'dog', exact: false })
-
-    t.equal(result3.count, 2)
-    t.equal(result4.count, 3)
-
-    // Typo-tolerant search
-    const result5 = await search(db, { term: 'fx', tolerance: 1 })
-    const result6 = await search(db, { term: 'dg', tolerance: 2 })
-
-    t.equal(result5.count, 2)
-    t.equal(result6.count, 4)
-
-    // Long string search (Tests for https://github.com/OramaSearch/orama/issues/159 )
-    const result7 = await search(db, { term: 'They are the best' })
-    const result8 = await search(db, { term: 'Foxes are nice animals' })
-
-    t.equal(result7.count, 2)
-    t.equal(result8.count, 2)
-  })
-
-  t.test('should correctly search without term', async t => {
-    t.plan(4)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-      components: {
-        tokenizer: {
-          stemming: true,
-        },
-      },
-    })
-
-    const docs = [
-      { id: '0', quote: 'the quick, brown fox jumps over the lazy dog. What a fox!', author: 'John Doe' },
-      { id: '1', quote: 'Foxes are nice animals. But I prefer having a dog.', author: 'John Doe' },
-      { id: '2', quote: 'I like dogs. They are the best.', author: 'Jane Doe' },
-    ]
-
-    await insert(db, docs[0])
-    await insert(db, docs[1])
-    await insert(db, docs[2])
-
-    // Exact search
-    const result1 = await search(db, { exact: false })
-    const result2 = await search(db, { exact: true })
-
-    t.equal(result1.count, 3)
-    t.equal(result2.count, 3)
-    t.strictSame(
-      result1.hits.sort((a, b) => a.id.localeCompare(b.id)).map(h => h.document),
-      docs,
-    )
-    t.strictSame(
-      result1.hits.sort((a, b) => a.id.localeCompare(b.id)).map(h => h.document),
-      docs,
-    )
-  })
-
-  t.test('should correctly search for data returning doc including with unindexed keys', async t => {
-    t.plan(4)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-      components: {
-        tokenizer: { language: 'english', stemming: false },
-      },
-    })
-
-    const documentWithUnindexedField = {
-      quote: 'I like cats. They are the best.',
-      author: 'Jane Doe',
-      unindexedField: 'unindexedValue',
-    }
-    const documentWithNestedUnindexedField = {
-      quote: 'Foxes are nice animals. But I prefer having a dog.',
-      author: 'John Doe',
-      nested: { unindexedNestedField: 'unindexedNestedValue' },
-    }
-
-    await insert(db, documentWithNestedUnindexedField)
-    await insert(db, documentWithUnindexedField)
-
-    const result1 = await search(db, { term: 'They are the best' })
-    const result2 = await search(db, { term: 'Foxes are nice animals' })
-
-    t.equal(result1.count, 1)
-    t.equal(result2.count, 1)
-    t.same(result1.hits[0].document, documentWithUnindexedField)
-    t.same(result2.hits[0].document, documentWithNestedUnindexedField)
-  })
-
-  t.test('should not found any doc if searching by unindexed field value', async t => {
-    t.plan(2)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-    })
-
-    await insert(db, {
-      quote: 'I like dogs. They are the best.',
-      author: 'Jane Doe',
-      nested: { unindexedNestedField: 'unindexedNestedValue' },
-    })
-
-    await insert(db, {
-      quote: 'I like cats. They are the best.',
-      author: 'Jane Doe',
-      unindexedField: 'unindexedValue',
-    })
-
-    const result1 = await search(db, { term: 'unindexedNestedValue' })
-    const result2 = await search(db, { term: 'unindexedValue' })
-
-    t.equal(result1.count, 0)
-    t.equal(result2.count, 0)
-  })
 
   t.test('should correctly insert and retrieve data', async t => {
     t.plan(4)
@@ -388,58 +234,6 @@ t.test('orama', t => {
     t.equal(ex1Search.count, 1)
     t.type(ex1Search.elapsed.raw, 'number')
     t.equal(ex1Search.hits[0].document.example, 'The quick, brown, fox')
-  })
-
-  t.test('should correctly paginate results', async t => {
-    t.plan(9)
-
-    const db = await create({
-      schema: {
-        animal: 'string',
-      },
-    })
-
-    await insert(db, { id: '0', animal: 'Quick brown fox' })
-    await insert(db, { id: '1', animal: 'Lazy dog' })
-    await insert(db, { id: '2', animal: 'Jumping penguin' })
-    await insert(db, { id: '3', animal: 'Fast chicken' })
-    await insert(db, { id: '4', animal: 'Fabolous ducks' })
-    await insert(db, { id: '5', animal: 'Fantastic horse' })
-
-    const search1 = await search(db, { term: 'f', limit: 1, offset: 0 })
-    const search2 = await search(db, { term: 'f', limit: 1, offset: 1 })
-    const search3 = await search(db, { term: 'f', limit: 1, offset: 2 })
-    const search4 = await search(db, { term: 'f', limit: 2, offset: 2 })
-
-    t.equal(search1.count, 4)
-    t.equal(search1.hits[0].document.animal, 'Fast chicken')
-
-    t.equal(search2.count, 4)
-    t.equal(search2.hits[0].document.animal, 'Fabolous ducks')
-
-    t.equal(search3.count, 4)
-    t.equal(search3.hits[0].document.animal, 'Fantastic horse')
-
-    t.equal(search4.count, 4)
-    t.equal(search4.hits[0].document.animal, 'Fantastic horse')
-    t.equal(search4.hits[1].document.animal, 'Quick brown fox')
-  })
-
-  t.test('should throw an error when searching in non-existing indices', async t => {
-    t.plan(1)
-
-    const db = await create({ schema: { foo: 'string', baz: 'string' } })
-
-    await t.rejects(
-      () =>
-        search(db, {
-          term: 'foo',
-          properties: ['bar'],
-        }),
-      {
-        code: 'UNKNOWN_INDEX',
-      },
-    )
   })
 
   t.test('should be able to insert documens with non-searchable fields', async t => {
@@ -474,223 +268,6 @@ t.test('orama', t => {
 
     t.equal(searchResult.count, 1)
     t.equal(searchResult.hits[0].document.author, 'Frank Zappa')
-  })
-
-  t.test('should exact match', async t => {
-    t.plan(4)
-
-    const db = await create({
-      schema: {
-        author: 'string',
-        quote: 'string',
-      },
-    })
-
-    await insert(db, {
-      quote: 'Be yourself; everyone else is already taken.',
-      author: 'Oscar Wilde',
-    })
-
-    const partialSearch = await search(db, {
-      term: 'alr',
-      exact: true,
-    })
-
-    t.equal(partialSearch.count, 0)
-
-    const exactSearch = await search(db, {
-      term: 'already',
-      exact: true,
-    })
-
-    t.equal(exactSearch.count, 1)
-    t.equal(exactSearch.hits[0].document.quote, 'Be yourself; everyone else is already taken.')
-    t.equal(exactSearch.hits[0].document.author, 'Oscar Wilde')
-  })
-
-  t.test("Shouldn't tolerate typos", async t => {
-    t.plan(1)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-    })
-
-    await insert(db, {
-      quote:
-        'Absolutely captivating creatures, seahorses seem like a product of myth and imagination rather than of nature.',
-      author: 'Sara A. Lourie',
-    })
-
-    const searchResult = await search(db, {
-      term: 'seahrse',
-      tolerance: 0,
-    })
-
-    t.equal(searchResult.count, 0)
-  })
-
-  t.test('should tolerate typos', async t => {
-    t.plan(2)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: 'string',
-      },
-    })
-
-    await insert(db, {
-      quote:
-        'Absolutely captivating creatures, seahorses seem like a product of myth and imagination rather than of nature.',
-      author: 'Sara A. Lourie',
-    })
-
-    await insert(db, {
-      quote: 'Seahorses look mythical, like dragons, but these magnificent shy creatures are real.',
-      author: 'Jennifer Keats Curtis',
-    })
-
-    const tolerantSearch = await search(db, {
-      term: 'seahrse',
-      tolerance: 2,
-    })
-
-    t.equal(tolerantSearch.count, 2)
-
-    const moreTolerantSearch = await search(db, {
-      term: 'sahrse',
-      tolerance: 5,
-    })
-
-    t.equal(moreTolerantSearch.count, 2)
-  })
-
-  t.test('should support nested properties', async t => {
-    t.plan(4)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: {
-          name: 'string',
-          surname: 'string',
-        },
-      },
-    })
-
-    await insert(db, {
-      quote: 'Harry Potter, the boy who lived, come to die. Avada kedavra.',
-      author: {
-        name: 'Tom',
-        surname: 'Riddle',
-      },
-    })
-
-    await insert(db, {
-      quote: 'I am Homer Simpson.',
-      author: {
-        name: 'Homer',
-        surname: 'Simpson',
-      },
-    })
-
-    const resultAuthorSurname = await search(db, {
-      term: 'Riddle',
-      properties: ['author.surname'],
-    })
-
-    const resultAuthorName = await search(db, {
-      term: 'Riddle',
-      properties: ['author.name'],
-    })
-
-    const resultSimpsonQuote = await search(db, {
-      term: 'Homer',
-      properties: ['quote'],
-    })
-
-    const resultSimpsonAuthorName = await search(db, {
-      term: 'Homer',
-      properties: ['author.name'],
-    })
-
-    t.equal(resultSimpsonAuthorName.count, 1)
-    t.equal(resultSimpsonQuote.count, 1)
-    t.equal(resultAuthorSurname.count, 1)
-    t.equal(resultAuthorName.count, 0)
-  })
-
-  t.test('should support multiple nested properties', async t => {
-    t.plan(3)
-
-    const db = await create({
-      schema: {
-        quote: 'string',
-        author: {
-          name: 'string',
-          surname: 'string',
-        },
-        tag: {
-          name: 'string',
-          description: 'string',
-        },
-      },
-    })
-
-    await insert(db, {
-      quote: 'Be yourself; everyone else is already taken.',
-      author: {
-        name: 'Oscar',
-        surname: 'Wild',
-      },
-      tag: {
-        name: 'inspirational',
-        description: 'Inspirational quotes',
-      },
-    })
-
-    await insert(db, {
-      quote: 'So many books, so little time.',
-      author: {
-        name: 'Frank',
-        surname: 'Zappa',
-      },
-      tag: {
-        name: 'books',
-        description: 'Quotes about books',
-      },
-    })
-
-    await insert(db, {
-      quote: 'A room without books is like a body without a soul.',
-      author: {
-        name: 'Marcus',
-        surname: 'Tullius Cicero',
-      },
-      tag: {
-        name: 'books',
-        description: 'Quotes about books',
-      },
-    })
-
-    const resultAuthor = await search(db, {
-      term: 'Oscar',
-    })
-
-    const resultTag = await search(db, {
-      term: 'books',
-    })
-
-    const resultQuotes = await search(db, {
-      term: 'quotes',
-    })
-
-    t.equal(resultAuthor.count, 1)
-    t.equal(resultTag.count, 2)
-    t.equal(resultQuotes.count, 3)
   })
 
   t.test('should suport batch insert of documents', async t => {
